+++ src/collective/registration/content/__init__.py
@@ -1,1 +1,0 @@
-from collective.registration.content import fgselectionfield